requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
pymupdf>=1.23.0
python-docx>=1.1.0
sentence-transformers>=2.2.2
openai>=0.28.0
requests>=2.31.0
numpy>=1.24.0

